Update include to "curr.h" and change at_splat argument to float

+++ src/fl/corkscrew.h
@@ -139,7 +139,8 @@ class Corkscrew {
     vec2f at_interp(float i) const;
     // This is a splatted pixel. This is will look way better than
     // using at(), because it uses 2x2 neighboor sampling.
-    Tile2x2_u8 at_splat(uint16_t i) const;
+    Tile2x2_u8 at_splat(float i) const;
+
     size_t size() const;
 
     iterator begin() { return mState.begin(); }
